pyproject.toml
[project]
name = "stardew-discord-bot"
version = "0.1.0"
description = "Stardew Helper – a Discord bot for Stardew Valley (slash commands, wiki links, fun facts)."
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [
  { name = "Stardew Helper Team" }
]
keywords = ["discord", "stardew valley", "bot", "slash commands", "i18n"]
classifiers = [
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.12",
  "License :: OSI Approved :: MIT License",
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "Topic :: Games/Entertainment",
  "Topic :: Internet",
]
dependencies = [
  "discord.py>=2.3.2",
  "pydantic-settings>=2.2.1",
  "python-dotenv>=1.0.1",
  "httpx>=0.27.0",
  "cachetools>=5.3.3",
]

[project.optional-dependencies]
dev = [
  "ruff==0.3.7",
  "mypy==1.10.0",
  "pytest==8.2.0",
  "pytest-asyncio==0.23.6",
  "types-python-dateutil",
  "types-requests",
  "types-toml",
]

[project.scripts]
stardew-bot = "stardew_bot.__main__:main"

[tool.ruff]
line-length = 100
target-version = "py312"
lint.select = [
  "E",
  "F",
  "I",
  "B",
  "UP",
  "DTZ",
  "ASYNC",
  "S",
]
lint.ignore = [
  "S311",
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101"]

[tool.ruff.lint.isort]
known-first-party = ["stardew_bot"]

[tool.mypy]
python_version = "3.12"
warn_unused_configs = true
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
strict_equality = true
ignore_missing_imports = true
exclude = ["tests/fixtures"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"

[tool.hatch.build]
include = ["src/stardew_bot/locales/*.json"]

[tool.hatch.build.targets.wheel]
packages = ["src/stardew_bot"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
